+++ tests/tests/pe_bliss2/directories/relocation_entry_tests.cpp
@@ -109,13 +109,13 @@ TEST(RelocationEntryTests, ApplyToTest)
 		(void)entry.apply_to(0u, 0u, {});
 	}, relocation_entry_errc::relocation_param_is_absent);
 
-	entry.get_param().emplace(0xedabu);
+	entry.get_param().emplace(static_cast<std::uint16_t>(0xedabu));
 	static constexpr std::uint16_t value16_2 = 0xe727u;
 	static constexpr std::uint64_t image_base_difference_2 = 0xFFFFFFFF21040000ull;
 	EXPECT_EQ(entry.apply_to(value16_2, image_base_difference_2, {}),
 		0x082cu);
 	
-	entry.get_param().emplace(0x5dabu);
+	entry.get_param().emplace(static_cast<std::uint16_t>(0x5dabu));
 	EXPECT_EQ(entry.apply_to(0xe727u, image_base_difference_2, {}),
 		0x082bu);
 }
